完成kd树最近邻搜索分类

+++ src/DataInput/DisMap.java
@@ -3,16 +3,16 @@ package DataInput;
 /**
  * 用于存储距离的数据格式*/
 public class DisMap {
-	int distance;
+	double distance;
 	int lable;
-	public DisMap(int d,int l){
+	public DisMap(double d,int l){
 		this.distance = d;
 		this.lable = l;
 	}
 	public double getDistance() {
 		return distance;
 	}
-	public void setDistance(int distance) {
+	public void setDistance(double distance) {
 		this.distance = distance;
 	}
 	public int getLable() {
